requirements.txt
amqp==2.3.2
billiard==3.5.0.4
bokeh==0.13.0
celery==4.2.1
confusable-homoglyphs==3.2.0
Django==2.0.2
django-cors-headers==2.4.0
django-extra-views==0.12.0
django-pure-pagination==0.3.0
django-redis-cache==1.7.1
django-registration==3.0
django-registration-redux==2.5
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
numpy==1.15.2
packaging==18.0
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.7.1
pyparsing==2.2.2
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.3
PyYAML==3.13
redis==2.10.6
six==1.11.0
tornado==5.1.1
vine==1.1.4
